# Conflicts:
#	sdk/cosmos/azure_data_cosmos/CHANGELOG.md
#	sdk/cosmos/azure_data_cosmos/build.rs
#	sdk/cosmos/azure_data_cosmos/src/constants.rs
#	sdk/cosmos/azure_data_cosmos_driver/CHANGELOG.md
#	sdk/cosmos/azure_data_cosmos_driver/Cargo.toml
#	sdk/cosmos/azure_data_cosmos_driver/build.rs
#	sdk/cosmos/azure_data_cosmos_driver/docs/GATEWAY_20_SPEC.md
#	sdk/cosmos/azure_data_cosmos_driver/src/driver/pipeline/operation_pipeline.rs
#	sdk/cosmos/azure_data_cosmos_driver/src/driver/transport/transport_pipeline.rs
#	sdk/cosmos/azure_data_cosmos_driver/src/error/cosmos_status.rs
#	sdk/cosmos/azure_data_cosmos_driver/src/fault_injection/fault_injecting_factory.rs
#	sdk/cosmos/azure_data_cosmos_driver/src/options/connection_pool.rs

+++ sdk/cosmos/azure_data_cosmos_driver/src/driver/transport/rntbd/status.rs
@@ -18,7 +18,7 @@ pub(crate) fn map_rntbd_status_to_cosmos_status(
     let status = StatusCode::from(http_status as u16);
     let mut cosmos_status = CosmosStatus::new(status);
     if let Some(sub_status) = sub_status {
-        cosmos_status = cosmos_status.with_sub_status(sub_status);
+        cosmos_status = cosmos_status.with_sub_status(sub_status as u16);
     }
     cosmos_status
 }
